Class file relocation based on functionality
Move the package to reorganise the class in term of functionality and not layers. This is for better expansion in future. So we will maintain the O (Open for extension, Closed for modification) of the SOLID principle
