+++ src/main/java/expense_tally/expense_manager/ExpenseReport.java
@@ -1,4 +1,4 @@
-package expense_tally.persistence;
+package expense_tally.expense_manager;
 
 import java.util.Objects;
 import java.util.StringJoiner;
