+++ src/main/java/expense_tally/expense_manager/ExpenseTransactionDao.java
@@ -1,7 +1,7 @@
-package expense_tally.persistence;
+package expense_tally.expense_manager;
 
-import expense_tally.model.ExpenseManager.ExpenseManagerMapKey;
-import expense_tally.model.ExpenseManager.ExpenseManagerTransaction;
+import expense_tally.expense_manager.model.ExpenseManagerMapKey;
+import expense_tally.expense_manager.model.ExpenseManagerTransaction;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -29,11 +29,11 @@ public class ExpenseTransactionDao {
 
     /**
      * Returns all the mappings between the content of the customised key and the list of
-     * {@link expense_tally.model.ExpenseManager.ExpenseManagerTransaction}.
+     * {@link ExpenseManagerTransaction}.
      * @return all the mapping between the content of the customised key and the list of
-     * {@link expense_tally.model.ExpenseManager.ExpenseManagerTransaction}.
+     * {@link ExpenseManagerTransaction}.
      * @throws SQLException when there is an error accessing the database
-     * @see expense_tally.model.ExpenseManager.ExpenseManagerMapKey
+     * @see ExpenseManagerMapKey
      */
     public Map<ExpenseManagerMapKey, List<ExpenseManagerTransaction>> getAllExpenseTransactions() throws SQLException {
         Map<Double, List<ExpenseManagerTransaction>> expenseTransactionMap = new HashMap();
